pyproject.toml
[tool.poetry]
name = "bob the bot (local)"
version = "0.1.0"
description = ""
authors = ["creapy_01 <creapy01.yt@gmail.com>"]
readme = "README.md"
package-mode = false


[tool.poetry.dependencies]
python = "^3.10"
discord-py = "^2.2.2"
python-dotenv = "^1.0.0"



[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
